requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.45.1
frozenlist==1.4.0
greenlet==3.0.1
idna==3.6
kiwisolver==1.4.5
magic-filter==1.0.12
matplotlib==3.8.2
multidict==6.0.4
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
pyqtgraph==0.13.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
websockets==12.0
yarl==1.9.4
